Refactor workflow execution modal and improve trigger selection logic
- Replaced `WorkflowEventModal` with `WorkflowExecuteModal` for a more flexible and structured execution flow.
- Enhanced trigger selection with support for manual, alert-based, and scheduled triggers.
- Updated related entity queries to include `networkMode: always` for consistent data fetching.
- Adjusted `i18n` translation keys for accuracy and consistency.
- Removed unused `workflow_event_modal.tsx` file.
- Updated component imports and services injection in the application context.
