+++ src/platform/plugins/shared/workflows_management/common/translations.ts
@@ -10,15 +10,33 @@
 import { i18n } from '@kbn/i18n';
 
 export const SEARCH_WORKFLOWS = i18n.translate(
-  'xpack.workflowsManagement.workflows.workflowsList.searchAriaLabel',
+  'xpack.workflowsManagement.workflowsList.searchAriaLabel',
   {
     defaultMessage: 'Search workflows',
   }
 );
 
 export const SEARCH_PLACEHOLDER = i18n.translate(
-  'xpack.workflowsManagement.workflows.workflowsList.searchPlaceholder',
+  'plugins.workflowsManagement.workflowsList.searchPlaceholder',
   {
     defaultMessage: 'Workflow name/description',
   }
 );
+
+export const MANUAL_TRIGGERS_DESCRIPTIONS: Record<string, string> = {
+  manual: i18n.translate(
+    'plugins.workflowsManagement.workflowsExecution.manualTriggerDescription',
+    {
+      defaultMessage: 'Provide values based on workflow definition.',
+    }
+  ),
+  alert: i18n.translate('plugins.workflowsManagement.workflowsExecution.alertTriggerDescription', {
+    defaultMessage: 'Select an alert to trigger workflow',
+  }),
+  scheduled: i18n.translate(
+    'plugins.workflowsManagement.workflowsExecution.scheduledTriggerDescription',
+    {
+      defaultMessage: 'Select a schedule to trigger workflow',
+    }
+  ),
+};
