+++ src/platform/plugins/shared/workflows_management/public/entities/workflows/model/use_workflow_execution.ts
@@ -7,16 +7,17 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { useQuery } from '@tanstack/react-query';
-import type { WorkflowDetailDto } from '@kbn/workflows';
 import { useKibana } from '@kbn/kibana-react-plugin/public';
+import type { WorkflowExecutionDto } from '@kbn/workflows';
+import { useQuery } from '@tanstack/react-query';
 
-export function useWorkflowDetail(id: string | null) {
+export function useWorkflowExecution(workflowExecutionId: string | null) {
   const { http } = useKibana().services;
 
-  return useQuery<WorkflowDetailDto>({
-    enabled: !!id,
-    queryKey: ['workflows', id],
-    queryFn: () => http!.get(`/api/workflows/${id}`),
+  return useQuery<WorkflowExecutionDto>({
+    networkMode: 'always',
+    queryKey: ['stepExecutions', workflowExecutionId],
+    queryFn: () => http!.get(`/api/workflowExecutions/${workflowExecutionId}`),
+    enabled: workflowExecutionId !== null,
   });
 }
